# docs/openapi.yaml
openapi: 3.0.3
info:
  title: My Social Networks API
  version: 1.0.0
  description: |
    API REST du nouveau service d'événements et de groupes de Facebook (TP Node.js / Express / MongoDB).

    **Authentification** : `POST /auth/register` ou `POST /auth/login` renvoient un token JWT (valable 24h)
    à envoyer dans le header `Authorization: Bearer <token>`.

    **Format des erreurs** : toutes les erreurs ont la forme `{ code, message, errors }`, où `errors`
    est la liste des problèmes à corriger (ex : erreurs de validation).

    **Listes** : les listes sont paginées avec `?page=1&limit=20` (100 maximum) et renvoient
    `{ data, pagination: { page, limit, total, pages } }`.

    **Visibilité** :
    - un groupe *secret* ou un événement *privé* répond `404` à ceux qui n'en font pas partie (il « n'existe pas » pour eux) ;
    - un groupe *privé* est visible (nom, description) mais ses membres, son fil et ses événements sont réservés aux membres (`403`).
  contact:
    name: Saad Lagzouli
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Développement
...
security:
  - bearerAuth: []
paths:
...
  /events/{id}/carpools:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getCarpools
      tags: [Carpooling]
      summary: Trajets proposés (participant)
      description: '`409` si `settings.carpooling` est désactivé.'
      responses:
        '200':
          description: Liste paginée, triée par heure de départ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CarpoolList' }
    post:
      operationId: createCarpool
      tags: [Carpooling]
      summary: Proposer un trajet en voiture (participant)
      description: Un trajet par conducteur et par événement.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CarpoolInput' }
      responses:
        '201':
          description: Trajet créé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Carpool' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /carpools/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getCarpoolById
      tags: [Carpooling]
      summary: Détail d'un trajet
      responses:
        '200':
          description: Trajet avec `seats_left`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Carpool' }
    patch:
      operationId: updateCarpool
      tags: [Carpooling]
      summary: Modifier son trajet (conducteur)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CarpoolInput' }
      responses:
        '200': { description: Trajet modifié }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      operationId: deleteCarpool
      tags: [Carpooling]
      summary: Supprimer un trajet (conducteur ou organisateur)
      responses:
        '200': { description: Trajet supprimé }
        '403': { $ref: '#/components/responses/Forbidden' }
  /carpools/{id}/passengers:
    post:
      operationId: joinCarpool
      tags: [Carpooling]
      summary: Réserver une place (participant)
      description: Réservation atomique, `409` s'il n'y a plus de place.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Place réservée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Carpool' }
        '409': { $ref: '#/components/responses/Conflict' }
  /carpools/{id}/passengers/{userId}:
    delete:
      operationId: leaveCarpool
      tags: [Carpooling]
      summary: Annuler une réservation (passager ou conducteur)
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200': { description: Réservation annulée }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, example: 6702f1c2a1b2c3d4e5f60718 }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Données invalides
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 400, message: Bad Request, errors: ['end_date doit être après start_date'] }
...
    Forbidden:
      description: Action non autorisée pour cet utilisateur
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 403, message: Forbidden, errors: ["action réservée aux organisateurs de l'événement"] }
    NotFound:
      description: Ressource introuvable (ou invisible pour cet utilisateur)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 404, message: Event Not Found }
    Conflict:
      description: Conflit avec l'état actuel (doublon, plus de place, règle métier)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 409, message: Conflict, errors: ["une personne ne peut obtenir qu'un seul billet par événement"] }
  schemas:
    Error:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        errors: { type: array, items: { type: string } }
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        pages: { type: integer, example: 3 }
    ObjectId:
      type: string
      example: 6702f1c2a1b2c3d4e5f60718
...
    UserPublic:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        firstname: { type: string, example: Saad }
        lastname: { type: string, example: Lagzouli }
        avatar: { type: string, format: uri, nullable: true }
        bio: { type: string, nullable: true }
...
    CarpoolInput:
      type: object
      required: [departure_place, departure_time, price, seats, max_detour_minutes]
      properties:
        departure_place: { type: string, maxLength: 300, example: 'Gare de Lyon, Paris' }
        departure_time: { type: string, format: date-time, example: '2026-11-14T17:30:00Z' }
        price: { type: number, minimum: 0, maximum: 1000, example: 5 }
        seats: { type: integer, minimum: 1, maximum: 8, example: 3 }
        max_detour_minutes: { type: integer, minimum: 0, maximum: 180, description: Temps maximum d'écart sur le trajet, example: 30 }
    Carpool:
      allOf:
        - $ref: '#/components/schemas/CarpoolInput'
        - type: object
          properties:
            _id: { $ref: '#/components/schemas/ObjectId' }
            event: { $ref: '#/components/schemas/ObjectId' }
            driver: { $ref: '#/components/schemas/UserPublic' }
            passengers: { type: array, items: { $ref: '#/components/schemas/UserPublic' } }
            seats_left: { type: integer }
    CarpoolList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Carpool' } }
        pagination: { $ref: '#/components/schemas/Pagination' }